+++ compiler/matchinvoke.cpp
@@ -11,7 +11,7 @@ void initializePatternEnv(EnvPtr patternEnv, llvm::ArrayRef<PatternVar> pvars,
     vector<PatternCellPtr> &cells, vector<MultiPatternCellPtr> &multiCells)
 {
     
-    for (unsigned i = 0; i < pvars.size(); ++i) {
+    for (size_t i = 0; i < pvars.size(); ++i) {
         if (pvars[i].isMulti) {
             MultiPatternCellPtr multiCell = new MultiPatternCell(NULL);
             multiCells.push_back(multiCell);
@@ -44,7 +44,7 @@ static void initializePatterns(OverloadPtr x)
     x->callablePattern = evaluateOnePattern(x->target, patternEnv);
 
     llvm::ArrayRef<FormalArgPtr> formalArgs = code->formalArgs;
-    for (unsigned i = 0; i < formalArgs.size(); ++i) {
+    for (size_t i = 0; i < formalArgs.size(); ++i) {
         FormalArgPtr y = formalArgs[i];
         PatternPtr pattern;
         if (y->type.ptr()) {
@@ -66,7 +66,7 @@ static void initializePatterns(OverloadPtr x)
 static void resetPatterns(OverloadPtr x)
 {
     assert(x->cells.size() == x->multiCells.size());
-    for (unsigned i = 0; i < x->cells.size(); ++i) {
+    for (size_t i = 0; i < x->cells.size(); ++i) {
         if (x->cells[i].ptr()) {
             x->cells[i]->obj = NULL;
             assert(!x->multiCells[i]);
@@ -114,8 +114,8 @@ MatchResultPtr matchInvoke(OverloadPtr overload,
             return new MatchArityError(code->formalArgs.size(), argsKey.size(), false);
     }
     llvm::ArrayRef<FormalArgPtr> formalArgs = code->formalArgs;
-    unsigned varArgSize = argsKey.size()-formalArgs.size()+1;
-    for (unsigned i = 0, j = 0; i < formalArgs.size(); ++i) {
+    size_t varArgSize = argsKey.size()-formalArgs.size()+1;
+    for (size_t i = 0, j = 0; i < formalArgs.size(); ++i) {
         FormalArgPtr x = formalArgs[i];
         if (x->varArg) {
             if (x->type.ptr()) {
@@ -140,7 +140,7 @@ MatchResultPtr matchInvoke(OverloadPtr overload,
     
     EnvPtr staticEnv = new Env(overload->env);
     llvm::ArrayRef<PatternVar> pvars = code->patternVars;
-    for (unsigned i = 0; i < pvars.size(); ++i) {
+    for (size_t i = 0; i < pvars.size(); ++i) {
         if (pvars[i].isMulti) {
             MultiStaticPtr ms = derefDeep(overload->multiCells[i].ptr());
             if (!ms)
@@ -166,7 +166,7 @@ MatchResultPtr matchInvoke(OverloadPtr overload,
         callable, argsKey
     );
     
-    for (unsigned i = 0, j = 0; i < formalArgs.size(); ++i) {
+    for (size_t i = 0, j = 0; i < formalArgs.size(); ++i) {
         FormalArgPtr x = formalArgs[i];
         if (x->varArg) {
             result->varArgName = x->name;
